Add code-shell class to show shell script commands
We usually add a dollar sign '$' to represent commands of shell script.
However, it's not convenient to copy these commands since you have to
remove the dollar sign manually.

The patch uses the ::before content: css to show the dollar sign so
it's not a part of the commands.

Refer to: http://jsfiddle.net/w3t5o2ee/